recreate initial migrations after successful connection to database + split IdentityModels.cs into ApplicationDbContext.cs and ApplicationUser.cs
