+++ LiveChat/Web.config
@@ -9,8 +9,23 @@
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
   <connectionStrings>
-    <add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-LiveChat-20180306114259.mdf;Initial Catalog=aspnet-LiveChat-20180306114259;Integrated Security=True"
+    <!--add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-LiveChat-20180306114259.mdf;Initial Catalog=aspnet-LiveChat-20180306114259;Integrated Security=True"
+      providerName="System.Data.SqlClient" /-->
+    <add name="DefaultConnection" 
+         connectionString=
+         "Data Source=WIN-AFU633JOFGJ\SQLEXPRESS;
+         Initial Catalog=LiveChat;
+         Integrated Security=True"
       providerName="System.Data.SqlClient" />
+    <!--add name="DefaultConnection" 
+         connectionString=
+         "Data Source=WIN-AFU633JOFGJ\SQLEXPRESS;
+         AttachDbFilename=|DataDirectory|\LiveChat.mdf;
+         Initial Catalog=aspnet-LiveChat-20180306114259;
+         Integrated Security=True"
+      providerName="System.Data.SqlClient" /-->
+    <!--add name="woodpecker" connectionString="Data Source=WIN-AFU633JOFGJ\SQLEXPRESS;AttachDbFilename=|DataDirectory|\LiveChat.mdf;Initial Catalog=LiveChat;Integrated Security=True;Network Library=dbmssocn" 
+         providerName="System.Data.SqlClient"/-->
   </connectionStrings>
   <appSettings>
     <add key="webpages:Version" value="3.0.0.0" />
@@ -22,6 +37,7 @@
     <authentication mode="None" />
     <compilation debug="true" targetFramework="4.6.1" />
     <httpRuntime targetFramework="4.6.1" />
+    <identity impersonate="false" />
   </system.web>
   <system.webServer>
     <modules>
